solve_captcha.py
#!/usr/bin/env python3
"""
Simple CAPTCHA solver for Amazon scraping
This is a placeholder - in production, you'd want to use a proper CAPTCHA solving service
"""

import sys
import requests
from PIL import Image
import io

def solve_captcha(captcha_url):
    """
    Placeholder CAPTCHA solver
    In production, integrate with services like:
    - 2captcha.com
    - Anti-Captcha.com
    - DeathByCaptcha
    """
    try:
        # Download the CAPTCHA image
        response = requests.get(captcha_url, timeout=10)
        response.raise_for_status()
        
        # For now, return a placeholder
        # In production, you would:
        # 1. Send image to CAPTCHA solving service
        # 2. Wait for solution
        # 3. Return the solved text
        
        print("CAPTCHA detected but no solver configured")
        print("Please integrate with a CAPTCHA solving service")
        return "PLACEHOLDER"
        
    except Exception as e:
        print(f"Error solving CAPTCHA: {e}")
        return "ERROR"

if __name__ == "__main__":
    if len(sys.argv) != 2:
        print("Usage: python solve_captcha.py <captcha_url>")
        sys.exit(1)
    
    captcha_url = sys.argv[1]
    solution = solve_captcha(captcha_url)
    print(solution) 