requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
pillow
lxml

